% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyflowchart.R
\name{applyflowchart}
\alias{applyflowchart}
\title{Annotate and label each reader mark based on a decision tree}
\usage{
applyflowchart(rawdata, PTcodes, PTCthreshold, includeA = T, includeC = T)
}
\arguments{
\item{rawdata}{The output datafile from the function readworkstation file.}

\item{PTcodes}{An R set object that contains the list of database ID codes that are included in the primary task.  If a reader mark is in this set, the ObsPTLesion variable in the output dataset will be coded as a 1. Else, it will be coded as 0. The same logic is applied to the classification of reference lesions and the outputted variable name is RefPTLesion.}

\item{PTCthreshold}{A numeric value used to determine the lower limit for any primary task confidence score to be considered in the analysis. The limit is applied as a greater than equal to. So specification of 1 will include all reader ROIs with a score greater than or equal to 1.}

\item{includeA}{A parameter, defaulted to True, that allows one to exclude type A reader marks. These are reader marks where they do not match any lesions in the reference set and the reader classifies them as a non-primary task lesion. Given the reader's confidence could be non-negigble, under most circumstations, these reader marks should be retained. Setting this to false will exclude these markings from all subsequent analyses.}

\item{includeC}{A parameter, defaulted to True, that allows one to exclude type C reader marks. These are reader marks where both the reader and reference localize the same non-primary task lesion.  Given the reader's confidence could be non-negigble, under most circumstations, these reader marks should be retained. Setting this to false will exclude these markings from all subsequent analyses.}
}
\value{
A new dataframe that appends several helper columns onto the input dataset.
}
\description{
This function normalizes the readerworkstation file (the data that is passed to this function) and prepares it for further analysis.  
The coding used to designate the primary task ROIs is a requried parameter. 
The includeA and includeC parameters define some special predetermined exclusions from the function. See the vignette for further elaboration of these parameters.
}
\examples{
applyflowchart(raw1, PTcodes=c("91"), PTCthreshold = 1, includeA=T,includeC=T)
}
