% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoanon.R
\docType{data}
\name{demoanon}
\alias{demoanon}
\title{Example anonomyization file}
\format{An object of class \code{data.frame} with 102 rows and 6 columns.}
\usage{
demoanon
}
\description{
Example of de-identification file
}
\keyword{datasets}
