% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\docType{data}
\name{demodata}
\alias{demodata}
\title{Example reader workstation file}
\format{An object of class \code{data.frame} with 15577 rows and 12 columns.}
\usage{
demodata
}
\description{
Example of a multi-reader, multi-case workstation datafile.
}
\keyword{datasets}
