% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedetections.R
\name{derivedetections}
\alias{derivedetections}
\title{Tabulate reader detections.}
\usage{
derivedetections(JAFROCfilename, wide = T)
}
\arguments{
\item{JAFROCfilename}{The is the excel file produced by the makeJAFROCfile function. This file is the source of most JAFROC analyses.}

\item{wide}{Logical flag that is used to indicate whether a wide (columns for various modalities) or tall (data rows are indexed by modality) dataset is returned. The default is a wide datafile.}
}
\value{
A dataframe with the detection results. For this determination, a detection is any localization where the primary task confidence is greater than or equal to the study minimum detection confidence as set in the \code{applyflowchart} function.
}
\description{
This function tabulates the number of true positives dections for each lesion at each modality.
The function reads the standard excel file that contains data to support JAFROC analyses.
}
\examples{
derivedetections("testJAFROC.xlsx")
derivedetections("testJAFROC.xlsx", wide=F)
}
