% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figuresave.R
\name{figuresave}
\alias{figuresave}
\title{Save package figure to PNG.}
\usage{
figuresave(filename, figobject, figheight = 15)
}
\arguments{
\item{filename}{The desired PNG filename, with extension}

\item{figobject}{The figure object (ggplot object) to be saved}

\item{figheight}{The desired heigh in CM. The default is set at 15 cm, which seems to work fine without further specification}
}
\description{
This is a simple wrapper function for \code{ggsave} to make a 4:3 ratio image that seems to produce 
saved PNGs that look reasonable from figures generated in this package.
}
\examples{
figuresave("PTCbyBins.PNG", plotPTC("testJAFROC.xlsx"))
}
