% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxConfidence.R
\name{getMaxConfidence}
\alias{getMaxConfidence}
\title{Internal function to get maximum confidence score.}
\usage{
getMaxConfidence(JAFROCfilename)
}
\arguments{
\item{JAFROCfilename}{The is the excel file produced by the makeJAFROCfile function. This file is the source of most JAFROC analyses.}
}
\value{
Returns the smallest maximum confidence score across all readers (and modalities).
}
\description{
This is a helper function to find the maximum confidence level reported by reader. 
This function avoids a bug where some readers will not report confidence to 100 for some lesions. 
This causes a for loop to error out in the \code{plotthresholds*} functions.
}
