% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeJAFROCfile.R
\name{makeJAFROCfile}
\alias{makeJAFROCfile}
\title{Produce JAFROC analysis storage file.}
\usage{
makeJAFROCfile(flowchartdata, outputfile)
}
\arguments{
\item{flowchartdata}{This is a dataframe produced by the function applyflowchart.}

\item{outputfile}{A text string for an excel file to be produced by the function. The file name can have the path and needs to end in .xlsx.}
}
\value{
This function does not have a return. It produces a saved output file for later use.
}
\description{
This is a key function in the JAFROC analysis paradigm. This is the datasource for the external package RJafroc. It also holds the data in a normalize function to support many analyses for this package.
}
\examples{
makeJAFROCfile(flowchartdata=raw2,outputfile="testJAFROC.xlsx")
}
