% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmctools.R
\docType{package}
\name{mrmctools}
\alias{mrmctools}
\alias{mrmctools-package}
\title{\code{mrmctools} package}
\description{
Package mrmctools to support multireader, multicase study designs
}
\details{
See the README on
\href{https://github.com/rickeycarter/mrmctools#readme}{GitHub}
}
