% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFOM.R
\name{plotFOM}
\alias{plotFOM}
\title{Plot Figure of Merit.}
\usage{
plotFOM(dataframe, themethod, thetreatment = "Treatment")
}
\arguments{
\item{dataframe}{This the dataframe that is the first list element returned by the prepareestimates function.  The array is processed outside this function to allow for labeling and formatting the modalities before the plot is generated.}

\item{themethod}{A character string that uses one of the following analysis methods: RRRC (random reader random case), FRRC (fixed reader, random case), RRFC (random reader, random case)}

\item{thetreatment}{A text string that indicates the name of the variable that pertains to the name of the modality. The JAFROC analysis will automatically use Treatment as this name. If for some reason the name is changed, this will allow that new name to be specified.}
}
\value{
\code{ggplot2} figure object
}
\description{
This function plots the Figure of Merits and 95\% CIs in a forest plot for each modality.
}
