% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPTC.R
\name{plotPTC}
\alias{plotPTC}
\title{Plot rating histogram.}
\usage{
plotPTC(JAFROCfilename)
}
\arguments{
\item{JAFROCfilename}{The is the excel file produced by the makeJAFROCfile function. This file is the source of most JAFROC analyses.}
}
\value{
\code{ggplot2} figure object.
}
\description{
This function plots a stacked histogram indicating the lesion localizations and non-lesion localizations by confidence rating. 
The function returns a ggplot object.
}
\examples{
plotPTC("testJAFROC.xlsx")
}
