% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPTCbins.R
\name{plotPTCbins}
\alias{plotPTCbins}
\title{Plot confidence scores by bin}
\usage{
plotPTCbins(flowdata, PTtext = "PT")
}
\arguments{
\item{flowdata}{A dataframe that has been returned by the applyflowchart function}

\item{PTtext}{A text string with the characters that is used describe the primary task. The function defaults to the generic text "PT".}
}
\description{
This function provides violin plots with jittered data for data according to the six classification bins for reader marks (see vignette). There is a simple parameter to change "PT" to something else, but the nature of x-axis design is that the replacement text needs to be brief.
}
