% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthreshold_modality.R
\name{plotthreshold_modality}
\alias{plotthreshold_modality}
\title{Plot reader sensitivity for a particular modality.}
\usage{
plotthreshold_modality(jafrocfile, modality = NA, modalitylabel = NA,
  includeGEE = T, readercolor = "#999999", geecolor = "#000000",
  getminmax = F)
}
\arguments{
\item{jafrocfile}{The is the excel file produced by the makeJAFROCfile function. This file is the source of most JAFROC analyses.}

\item{modality}{The internal variable name for the modality to be printed.}

\item{modalitylabel}{A character string to replace the variable name in the figure.}

\item{includeGEE}{Logical parameter to include the GEE line on the figure. Defaults to true.}

\item{readercolor}{Hexidecimal color code for the reader lines. Defaults to light grey.}

\item{geecolor}{Hexidecimal color code for the GEE line. Defaults to black.}

\item{getminmax}{A logical parameter to indicate a call to the \code{getMaxConfidence} function. If set to true, the figure will be truncated at this limit.}
}
\value{
\code{ggplot2} figure object.
}
\description{
This function produces a plot that shows how reader detections fall off in sensitivity as the sensitivity detection threshold increases.
}
