% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthreshold_reader.R
\name{plotthreshold_reader}
\alias{plotthreshold_reader}
\title{Plot reader sensitivity for a particular reader.}
\usage{
plotthreshold_reader(jafrocfile, readerid = "GEE", addFactor = F,
  flevels = NA, flabels = NA, getminmax = F)
}
\arguments{
\item{readerid}{Character string for the reader ID. The default is "GEE" which means the figure will only show the GEE lines by modality.}

\item{addFactor}{If set to true, the function will use the flevels and flabels to format the modalities}

\item{flevels}{Text object that contains the factor levels for formatted printing}

\item{flabels}{Text object that contains the factor labels for formatted printing}

\item{getminmax}{A logical parameter to indicate a call to the \code{getMaxConfidence} function. If set to true, the figure will be truncated at this limit.}

\item{JAFROCfilename}{The is the excel file produced by the makeJAFROCfile function. This file is the source of most JAFROC analyses.}
}
\value{
\code{ggplot2} figure object.
}
\description{
This function produces a plot that shows how reader detections vary by modality as the sensitivity detection threshold increases. Note, this function
takes a parameter that indicates the reader number. This needs to be a two character text string with a leader "0" if necessary ("01" for reader 1).
}
