% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareestimates.R
\name{prepareestimates}
\alias{prepareestimates}
\title{Preprocess raw JAFROC results.}
\usage{
prepareestimates(JAFROCresults, refname)
}
\arguments{
\item{JAFROCresults}{This is object produced by the package \code{RJafroc}. See the vingette for details.}

\item{refname}{Character string that has the internal value indicating the reference dose for the study.}
}
\value{
Array filled with results.
}
\description{
This function reads and prepares the result object generated from \code{RJafroc} for plotting.
}
