% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readworkstationfile.R
\name{readworkstationfile}
\alias{readworkstationfile}
\title{Read and normalize reader workstation file.}
\usage{
readworkstationfile(workstationdata, deidentificationdata, refname, caseindex,
  RefID, RefDataID, RefDetConf, RefCode, RefPTC, datasetname, ObserverPTC,
  ObserverCode, ObserverROIID, ReaderFullID, anonymizereader = T,
  expandcases = F)
}
\arguments{
\item{workstationdata}{A dataframe with the raw Discovery Workstation export file}

\item{deidentificationdata}{A dataframe with columns for patient ID a column with dataset IDs for each modality in the study}

\item{refname}{Text string with the name of the reference modality}

\item{caseindex}{Text string with the name of the patient ID variable in the deidentification file}

\item{RefID}{Text string with the variable name for the reference marking unique identifier in the workstation file}

\item{RefDataID}{Text string with the variable name for the reference dose dataset ID in the workstation file}

\item{RefDetConf}{Text string with the variable name for the confidence score related to the reference detection (not primary task)}

\item{RefCode}{Text string with the varaible name for diagnosis code assigned by the reference reader for the the lesion}

\item{RefPTC}{Text string with the varaible name for the primary task confidence (PTC).}

\item{datasetname}{Text string with the variable name for the reader's dataset name pertaining to the ROI being evaluated}

\item{ObserverPTC}{Text string with the variable name for the reader's assigned primary task confidence (PTC). This is the main confidence score used for the analysis.}

\item{ObserverCode}{Text string with the variable name for the reader's assigned diagnosis of the ROI.}

\item{ObserverROIID}{Text string for the variable name for the reader's unique ROI database ID.}

\item{ReaderFullID}{Text string with the username (typically) used to log into the study. This represents the starting point for the reader identification in the dataset.}

\item{anonymizereader}{A logical to replace the internal reader ID with an anonymized reader ID by default. Setting this to false will possibly identify the data in the analysis.}

\item{expandcases}{A logical to ensure that the joins that happen in the program include all combinations of readers and modalities. Generally this needs to remain as True.}
}
\value{
A dataframe with standardized names
}
\description{
This is the first function called when conducting a MRMC study using the Discovery Workstation computer system. 
The workstation has algorithm that match reader marks with reference ROIs. The output file is study-specific.
To help generalize the analysis and all of the functions that follow, this function renames the study-specific variables into 
standardized names.
}
\details{
The function requires the use of an anonymization file. This file links back the datasets (image files) back to doses and patient cases.
See the vignette for an example and elaboration of this file.
}
